Added CreateUserMenu class that handles all UI-related matters for the new server code setup.
This will allow us to more easily implement the new server code system across our applications. It also unifies the process and UI Flow for creating new users. This does utilize the ModalDialog class, however. So, any apps not using ModalDialog will not be able to use this at the moment.