Consolidated Logging Name Implementation
Put the logging name assignment into the ProfileData constructor. Included option logging name parameters on PlayerData.AddUser so it can be used with new versions of our apps that use the new server system and not break legacy apps that are not yet using it. Verified the changes are working in Recollect.