Added POST requests that can take a file download handler
This allows us to download files via a POST request, if needed. While this is not a common use case, we have an endpoint for module manifests that requires a body containing data to archive into a ZIP file. If Unity could send GET requests with a body, this wouldn't be an issue, but you can't.